requirements.txt
beautifulsoup4
fastapi
httpx
pydantic
Requests
sse_starlette
uvicorn
openai
python-dotenv